pyproject.toml
[tool.poetry]
name = "business-advisor"
version = "0.1.0"
description = "Agentic Business Advisory System"
authors = ["osmnclskn7@gmail.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
pydantic-settings = "^2.12.0"
pydantic = "^2.9.0"
pyyaml = "^6.0.3"
langchain-core = "^1.2.7"
langchain-openai = "^1.1.7"
langchain-anthropic = "^1.3.1"
langchain-google-genai = "^4.2.0"
tavily-python = "^0.7.21"
motor = "^3.7.1"
langgraph = "^1.0.7"
langgraph-checkpoint-mongodb = "^0.3.1"
redis = "^7.1.0"
black = "^26.1.0"
isort = "^7.0.0"
celery = "^5.6.2"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"